Update variable type

+++ arduino/software/src/main1.ino
@@ -4,10 +4,10 @@
 
 /* データの構造 */
 typedef struct struct_message {
-  int mode1;      // 送信モードを指定
-  int number[10]; // 送信するユニットを指定
-  int color[3];   // 送信する色を指定
-  int mode2;      // 遅延時間を指定
+  uint8_t mode1;      // 送信モードを指定
+  uint8_t number[10]; // 送信するユニットを指定
+  uint8_t color[3];   // 送信する色を指定
+  uint8_t mode2;      // 遅延時間を指定
 } struct_message;
 
 /* Global variables */
@@ -24,7 +24,7 @@ IPAddress gateway(IP_ADDRESS_GATEWAY);
 IPAddress subnet(IP_ADDRESS_SUBNET);
 
 // OSC
-const int port = OSC_PORT;
+const uint16_t port = OSC_PORT;
 
 /* データ送信時のコールバック関数 */
 void OnDataSent(uint8_t *mac_addr, uint8_t sendStatus) {
